rename /search/search to /search
